# Speed up (some) interpreters compilation