--- a/programs/copier/ClusterCopierApp.cpp
+++ b/programs/copier/ClusterCopierApp.cpp
@@ -2,6 +2,7 @@
 #include <Common/StatusFile.h>
 #include <Common/TerminalSize.h>
 #include <IO/ConnectionTimeoutsContext.h>
+#include <Interpreters/DatabaseCatalog.h>
 #include <Formats/registerFormats.h>
 #include <base/scope_guard_safe.h>
 #include <unistd.h>