--- a/programs/local/LocalServer.cpp
+++ b/programs/local/LocalServer.cpp
@@ -27,6 +27,7 @@
 #include <IO/ReadBufferFromString.h>
 #include <IO/WriteBufferFromFileDescriptor.h>
 #include <IO/UseSSL.h>
+#include <Parsers/ASTInsertQuery.h>
 #include <Parsers/IAST.h>
 #include <base/ErrorHandlers.h>
 #include <Functions/registerFunctions.h>