--- a/src/Columns/ColumnNullable_fwd.h
+++ b/src/Columns/ColumnNullable_fwd.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <Common/COW.h>
+
+namespace DB
+{
+
+class IColumn;
+using ColumnPtr = COW<IColumn>::Ptr;
+
+class ColumnNullable;
+ColumnPtr makeNullable(const ColumnPtr & column);
+
+}