--- a/src/Common/ThreadProfileEvents.cpp
+++ b/src/Common/ThreadProfileEvents.cpp
@@ -10,7 +10,7 @@
 #include <fstream>
 #include <optional>
 #include <sstream>
-#include <unordered_set>
+#include <unordered_map>
 
 #include <fcntl.h>
 #include <unistd.h>