--- a/src/Functions/nullIf.cpp
+++ b/src/Functions/nullIf.cpp
@@ -1,10 +1,9 @@
-#include <Functions/IFunction.h>
-#include <Functions/FunctionHelpers.h>
-#include <Functions/FunctionFactory.h>
-#include <DataTypes/DataTypesNumber.h>
-#include <DataTypes/DataTypeNullable.h>
+#include <Columns/ColumnNullable_fwd.h>
 #include <Core/ColumnNumbers.h>
-#include <Columns/ColumnNullable.h>
+#include <DataTypes/DataTypeNullable.h>
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionHelpers.h>
+#include <Functions/IFunction.h>
 
 
 namespace DB