--- a/src/Functions/nullIf.cpp
+++ b/src/Functions/nullIf.cpp
@@ -1,9 +1,10 @@
-#include <Columns/ColumnNullable_fwd.h>
-#include <Core/ColumnNumbers.h>
-#include <DataTypes/DataTypeNullable.h>
-#include <Functions/FunctionFactory.h>
-#include <Functions/FunctionHelpers.h>
 #include <Functions/IFunction.h>
+#include <Functions/FunctionHelpers.h>
+#include <Functions/FunctionFactory.h>
+#include <DataTypes/DataTypesNumber.h>
+#include <DataTypes/DataTypeNullable.h>
+#include <Core/ColumnNumbers.h>
+#include <Columns/ColumnNullable.h>
 
 
 namespace DB