Changelog category (leave one):
- Not for changelog (changelog entry is not required)


Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
...


Detailed description / Documentation draft:

This is a follow up to #31896 where I tried reducing CH compilation by reducing how many files needed to be rebuilt after modifying some headers.

In this PR instead I tried a different approach. Take a few files and verify why they are slow to build and try to improve it. The results:
* InterpreterSelectWithUnionQuery.cpp: 10.2 seconds to 6.5 seconds
* InterpreterSelectWithUnionQuery.cpp: 22.5 seconds to 22.1 seconds
* InterpreterInsertQuery.cpp: 7.1 seconds to 6.4 seconds

Some comments about this:
* I did these tests and measurements in a laptop with crazy frequency autoscaling, so YMMV.
* Was the effort worth it? If you consider the speed improvement no, but I got to familiarize myself with `-ftime-trace` and understand it's output and so on.
* There are a few files that affect a lot of compilation units that are pretty to build: `Fields.h`, context, settings, aggregations... Maybe in the future (once compiler and tooling is there) modules will help in some of those cases.
* Removing an include from a header often leads to dozens or more changes required in `cpp` files to introduce (where needed) its true dependencies which before were included transitively. Most of the changes are because of that.
* I'd say the most controversial changes would be the introduction of `SettingsEnums_fdw.h` containing only the enums and not the classes declarations and of `Interpreters/Loggers.h` to move the dependency on `Interpreters/SystemLog.h` (slow file) there.