@alexey-milovidov 2021-11-29
@Algunenano I always have a dream to collect `-ftime-trace` for every build and load it into ClickHouse.
If you can provide an example of the table structure and how to better upload this data into ClickHouse it should be very helpful. I did not check it by myself yet.