@Algunenano 2021-11-29
Generating them is super simple; just adding a flag to CFLAGS/CXXFLAGS will generate the same .o but also an additional .json with the tracing information. Sample file: [StorageSystemAsynchronousMetrics.cpp.json](https://github.com/ClickHouse/ClickHouse/files/7620543/StorageSystemAsynchronousMetrics.cpp.txt)

All I do is load individual json files to Chrome's about:tracing or https://ui.perfetto.dev/ (which is the same but with a better UI) as it uses Trace Event Format JSON Object Format. The tricky part (or rather time consuming) is to come up with the proper structure/tables to load that into ClickHouse and extract meaningful information.